requirements.txt
fastapi
uvicorn[standard]
python-multipart
soundfile
nvidia-riva-client

